For SVY-15312: As Multi SCM Jenkins plugin is now deprecated we cannot use multiple git repositories when configuring a freestyle job. That means that we will use git submodules - which have some limitations (like poll doesn't work properly - although ppl should use git post commit hook triggers/github triggers).
However Jenkins pipeline approach (what the file from this commit starts
to do) - instead of freestyle job would still allow this and it seems to
be the way to go but it is harder to use for non-UI ppl. So for now wiki
and sample remain freestyle jobs. Maybe in the future blue ocean for
Jenkins is much more advanced and easier to use as UI for pipeline...
Which is funny as we'd be back to jobs with UI (where we started :D).